1.dubbo对内开发 2.cpc计算开发

+++ cpct-service/src/main/java/com/zjtelcom/cpct/service/impl/report/MktCampaingReportServiceImpl.java
@@ -341,7 +341,7 @@ public class MktCampaingReportServiceImpl implements MktCampaingReportService {
             cityMap.put("name", sysArea.getName());
             int count = 0;
             for (MktCampaignDO mktCampaignDO : labelList) {
-                if (sysArea.getAreaId().equals(mktCampaignDO.getLanId())) {
+                if (sysArea.getAreaId().equals(mktCampaignDO.getLanId().intValue())) {
                     count++;
                 }
             }
